requirements.txt
flask==1.1.4
flask_login==0.5.0
flask_migrate==2.6.0
flask_security==3.0.0
flask_wtf==0.15.1
flask_sqlalchemy==2.5.1
email_validator
python-decouple
gunicorn==20.1.0
flask_script==2.0.6
requests==2.25.1
Flask-RESTful==0.3.9
flask-marshmallow==0.14.0
marshmallow-sqlalchemy
flask-fixtures==0.3.8
PyYAML==5.4.1
flask_httpauth==4.4.0
Flask-Principal==0.4.0
keystoneauth1==4.4.0
PyMySQL==1.0.2

